Try image dataset and debug
